test : MenuService 통합테스트 작성

+++ src/main/java/kitchenpos/domain/Product.java
@@ -10,9 +10,9 @@ public class Product {
     public Product() {
     }
 
-    public Product(String name, BigDecimal price) {
+    public Product(String name, Long price) {
         this.name = name;
-        this.price = price;
+        this.price = BigDecimal.valueOf(price);
     }
 
     public Long getId() {
